requirements.txt
Flask
gunicorn
keras.utils
keras.preprocessing
tensorflow-cpu